COW implemented. a new syscall added(sys_phy_page).
and again a bug hidden is found.
in page_insert, when the page need inserting is there,
we do not need to insert it again, but the permission
could be changed. my original statement is:
	*pte |= perm;
as you can see, it has the bug that when the *pte has PTE_W,
while perm does not. in this condition, the corresponding
result is not what we want. ;-/

Signed-off-by: Guanqun Lu <Guanqun.Lu@Gmail.com>