bug in serve_close fixed.
The bug is exposed when we use 'dup' function. As we know,
after dup, two references on the client side, one reference
on the file sytem server side. And 'close' is invoked, finally
'serve_close' is used, then in my original implementation,
the only reference on the fs server side is disposed at once,
without even knowing that other clients are still using it.
Otherwise, the following fs request to the server will just
abort in 'openfile_lookup', because the only reference is lost
on the server side, and it can not be accessed. Actually,
in this case, 'close' return a negative value showing that
the fileid is an invalid value, but in most cases, we ignore
the return value of 'close'. Sorry to mention that the world
is not perfect, even 'close' may fail... :-( Fortunately,
the bug now is fixed, and you can call 'close' as many times
as you want without caring about the return value. :-) World
perfect now? Hope so...

Signed-off-by: Guanqun Lu <guanqun.lu@gmail.com>